fix/gen: import paths after moving stuff around
