+++ test/runtime/java/app/test/sample-java.test.js
@@ -1,4 +1,4 @@
-const cds_test = require("../../../lib/cds-test");
+const cds_test = require("../../../../../lib/cds-test");
 
 describe("Java integration", () => {
   if (!/java/.test(process.env.CDS_ENV))
