+++ test/runtime/node/cds-test.test.js
@@ -1,13 +1,13 @@
 const cds = require ('@sap/cds/lib') // using cds/lib here to bypass @types
-const cds_test = require ('..')
+const cds_test = require ('../../..')
 const Books = 'sap.capire.bookshop.Books'
 
 describe('cds_test', ()=>{
 
-  const { expect, test } = cds_test().in(__dirname,'node-app')
+  const { expect, test } = cds_test().in(__dirname,'app')
 
   it('should have started the server correctly', () => {
-    const bookshop = cds.utils.path.resolve(__dirname,'node-app')
+    const bookshop = cds.utils.path.resolve(__dirname,'app')
     expect (cds.env._home) .to.equal (bookshop)
     expect (cds.root) .to.equal (bookshop)
     expect (test.server).to.exist
